pyproject.toml
[project]
name = "commonplace"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "pyyaml",
    "xdk",
]
